requirements.txt
streamlit
groq
pymongo
python-dotenv
motor
easyocr